fix: do not set channel if executablePath is provided
